+++ templates/Web/Projects/AngularDefault/src/app/app-shell/nav-bar/nav-bar.component.html
