+++ templates/Web/Projects/AngularDefault/src/app/app-shell/nav-bar/nav-bar.component.spec.ts
@@ -1,6 +1,6 @@
 ﻿import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { HeaderComponent } from './navbar.component';
+import { HeaderComponent } from './nav-bar.component';
 
 describe('HeaderComponent', () => {
   let component: HeaderComponent;
