+++ templates/Web/Projects/AngularDefault/src/app/app-shell/nav-bar/nav-bar.component.ts
@@ -2,9 +2,9 @@
 import { Location } from '@angular/common';
 
 @Component({
-  selector: 'app-header',
-  templateUrl: './navbar.component.html',
-  styleUrls: ['./navbar.component.css']
+  selector: "app-header",
+  templateUrl: "./nav-bar.component.html",
+  styleUrls: ["./nav-bar.component.css"]
 })
 export class HeaderComponent implements OnInit {
   skipLinkPath: string;
